Refactor project structure into modular packages
